+++ R/coneproj.fit.R
@@ -7,17 +7,21 @@
 coneproj.fit <- function(Dmat, dvec, Cmat, lb, ub, qp_pars){
   #----- Construct Cmat and bvec from lb and ub
 
-  # Get lb constraints
-  lbcons <- lb > -Inf
-  Amat <- Cmat[lbcons,,drop = F]
-  bvec <- lb[lbcons]
-  cmap <- which(lbcons)
+  if (NROW(Cmat) > 0){
+    # Get lb constraints
+    lbcons <- lb > -Inf
+    Amat <- Cmat[lbcons,,drop = F]
+    bvec <- lb[lbcons]
+    cmap <- which(lbcons)
 
-  # Get ub constraints
-  ubcons <- ub < Inf
-  Amat <- rbind(Amat, -Cmat[ubcons,])
-  bvec <- c(bvec, -ub[ubcons])
-  cmap <- c(cmap, which(ubcons))
+    # Get ub constraints
+    ubcons <- ub < Inf
+    Amat <- rbind(Amat, -Cmat[ubcons,])
+    bvec <- c(bvec, -ub[ubcons])
+    cmap <- c(cmap, which(ubcons))
+  } else {
+    stop("coneproj does not accept empty Cmat, use another solver")
+  }
 
   #----- Fit