+++ R/shapeConstr.factor.R
@@ -31,6 +31,10 @@ shapeConstr.factor <- function(x, shape, intercept = FALSE, ...) {
   chkc <- checkCmat(Cmat)
   if (length(chkc$redundant) > 0) Cmat <- Cmat[-chkc$redundant, , drop = F]
 
+  # Add bound attributes
+  attr(Cmat, "lb") <- rep(0, NROW(Cmat))
+  attr(Cmat, "ub") <- rep(Inf, NROW(Cmat))
+
   # Return
   Cmat
 }