+++ R/shapeConstr.ns.R
@@ -21,8 +21,14 @@ shapeConstr.ns <- function(x, shape, ...){
   qr.const <- qr(t(const))
   Cmat <- as.matrix((t(qr.qty(qr.const, t(Cmat))))[, -(1L:2L)])
 
-  # Return
+  # Constraining of NS can create some redundant constraints
   chkc <- checkCmat(Cmat)
   if (length(chkc$redundant) > 0) Cmat <- Cmat[-chkc$redundant, , drop = F]
+
+  # Add bound attributes
+  attr(Cmat, "lb") <- rep(0, NROW(Cmat))
+  attr(Cmat, "ub") <- rep(Inf, NROW(Cmat))
+
+  # Return
   Cmat
 }